Move util.Validate* to the only user
This removes the "github.com/go-playground/validator/v10"
dependency from pkg/client users, saving 0.66 MB.

Alternatively, this could be moved into some pkg/cli if more CLI
users are anticipated.
